couple of small logic tweaks

+++ src/db_download.sh
@@ -3,6 +3,11 @@
 echo "Downloading a dump of the database..."
 pg_dump -h $REMOTE_DB_HOST -p $REMOTE_DB_PORT -d $REMOTE_DB_NAME -U $REMOTE_DB_USER -W -Fc > ../latest.dump
 
+if [ $? -ne 0 ]
+    then
+        exit
+fi
+
 echo "Creating a new database locally..."
 psql -c "DROP DATABASE IF EXISTS local_prune;"
 psql -c "CREATE DATABASE local_prune;"
